Update form rendering to Bootstrap 5 (#3)
* Replace `.sr-only` class with `.visually-hidden`

Bootstrap 5 has switched the "screen reader" class from `.sr-only` to
`.visually-hidden`. Update the library so that form labels which are
set to be hidden are actually hidden while stile being available for
assistive technologies.

This change is not currently backwards compatible, but could easily be
made so by adding and extra check to renderers line #474 for the old
'sr-only' value.

* Remove `.form-group` and use `.mb-3`

Bootstrap 5 removes the `.form-group` class in preference of using the
utility class `.mb-3`, or another of the utility classes which fits the
use case better. Replace the former with the latter in the code.

This change in Bootstrap seems to be aiming to make form groups more
flexible - this code should not have a problem setting `.mb-3` since it
is only a default and can be overridden by using the `form_group_class`
parameter.

* Add `.form-label` class to labels

Bootstrap 5 introduced the `.form-label` class alongside removing the
`.form-group` class. Add `.form-label` to the label element if no other
class is already present.

* Update select and file widget classes

The file upload field in older versions of Bootstrap needed its own
special class. Bootstrap 5 changed that and allowed it to work with the
default `.form-control` class. Remove its special treatment.

The select widget used to work with the default `.form-control` class,
but was updated to use `.form-select`. Make an exception for choice
fields.

* Add the `.form-label` class to `render_label`

The `render_label` template tag needs to have the new form label class
provided to it. Since it is such a simple function instead of a complex
renderer just setting the class as the parameter default value does
what it needs.

* Implement inline forms with `.row`

Bootstrap 5 dropped the `.form-inline` class and suggests using `.row`
on the form or parent element and `.col-auto` on the "form group"
element. Implement this by checking on inline fields if the class value
is simply the default `.mb-3` and replacing it with `.col-auto` if it
is.

* Update docstrings on templatetag functions.

Make changes to a few of the templatetag functions' docstrings so that
the documentation is kept up to date with Bootstrap 5 friendly changes.